Move device_impl
