+++ tests/tt_metal/tt_metal/dispatch/random_program_fixture.hpp
@@ -6,7 +6,7 @@
 
 #include "command_queue_fixture.hpp"
 #include "env_lib.hpp"
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 #include <tt-metalium/host_api.hpp>
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/circular_buffer_constants.h>
