+++ tt-train/sources/ttml/core/ttnn_all_includes.hpp
@@ -15,7 +15,7 @@
 #include <hostdevcommon/common_values.hpp>                                                         // NOLINT
 #include <tt-metalium/base_types.hpp>                                                              // NOLINT
 #include <tt-metalium/bfloat16.hpp>                                                                // NOLINT
-#include <tt-metalium/device_impl.hpp>                                                             // NOLINT
+#include <tt-metalium/device.hpp>                                                                  // NOLINT
 #include <tt-metalium/distributed_context.hpp>                                                     // NOLINT
 #include <tt-metalium/host_api.hpp>                                                                // NOLINT
 #include <tt-metalium/math.hpp>                                                                    // NOLINT
