+++ tt_metal/CMakeLists.txt
@@ -52,7 +52,6 @@ target_sources(
             api/tt-metalium/data_format.hpp
             api/tt-metalium/data_types.hpp
             api/tt-metalium/device.hpp
-            api/tt-metalium/device_impl.hpp
             api/tt-metalium/device_pool.hpp
             api/tt-metalium/dispatch_core_common.hpp
             api/tt-metalium/distribution_spec.hpp
