+++ tt_metal/impl/lightmetal/lightmetal_replay_impl.cpp
@@ -15,7 +15,7 @@
 #include <tt-metalium/tt_metal.hpp>
 #include <tt-metalium/trace_buffer.hpp>
 #include <tt-metalium/command_queue.hpp>
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 #include "flatbuffer/base_types_from_flatbuffer.hpp"
 #include "flatbuffer/program_types_from_flatbuffer.hpp"
 #include "flatbuffer/buffer_types_from_flatbuffer.hpp"
