+++ tt_metal/programming_examples/matmul_single_core/matmul_single_core.cpp
@@ -9,7 +9,7 @@
 #include <tt-metalium/tilize_utils.hpp>
 #include <tt-metalium/command_queue.hpp>
 #include <matmul_common/bmm_op.hpp>
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 
 using namespace tt::constants;
 using namespace std;
