+++ tt_metal/programming_examples/vecadd_multi_core/vecadd_multi_core.cpp
@@ -8,7 +8,7 @@
 #include <tt-metalium/bfloat16.hpp>
 #include <tt-metalium/core_coord.hpp>
 #include <tt-metalium/host_api.hpp>
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 #include <tt-metalium/work_split.hpp>
 
 #include <cstddef>
