+++ tt_metal/programming_examples/vecadd_sharding/vecadd_sharding.cpp
@@ -10,7 +10,7 @@
 #include <tt-metalium/bfloat16.hpp>
 #include <tt-metalium/core_coord.hpp>
 #include <tt-metalium/host_api.hpp>
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 #include <tt-metalium/tt_metal.hpp>
 
 #include <cstddef>
