+++ ttnn/api/ttnn/run_operation.hpp
@@ -10,7 +10,7 @@
 #include "ttnn/operations/experimental/auto_format/auto_format.hpp"
 #include "ttnn/operation.hpp"
 #include "ttnn/common/queue_id.hpp"
-#include <tt-metalium/device_impl.hpp>
+#include <tt-metalium/device.hpp>
 #include <tt_stl/type_name.hpp>
 
 namespace tt::tt_metal {
